# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RoyaStar Backend API",
        "description": "LMS platform for managing students, teachers, schools, and curriculum.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "6535290507fbf3a0a4459475136f4a05",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered / School request pending"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "05a4834111390b8216b8a8cb1a1bc0ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "Phone number or email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token returned"
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/auth/send-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send OTP",
                "operationId": "25badcdb2b6c8843f9432bfd87a2d84a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent"
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/auth/verify-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify OTP",
                "operationId": "ec01eaf4af3c1dc827c8fb241183c74c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "maxLength": 6,
                                        "minLength": 6
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP verified"
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/auth/admin-login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Admin login",
                "operationId": "423c8b679c9ae90e40850c02d84896a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token returned"
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout (Bearer)",
                "operationId": "ee211fe2ca06511050c01942ccdb183c",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status_code": {
                        "type": "integer",
                        "example": 422
                    },
                    "message": {
                        "type": "string",
                        "example": "Something went wrong."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field": [
                                "Error message"
                            ]
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "RegisterRequest": {
                "required": [
                    "email",
                    "type"
                ],
                "properties": {
                    "name": {
                        "description": "Required for non-school types",
                        "type": "string"
                    },
                    "surname": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "description": "Required for non-school types",
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Required for non-school types",
                        "type": "string",
                        "minLength": 8
                    },
                    "password_confirmation": {
                        "description": "Must match password, required with password",
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "default": "student",
                        "enum": [
                            "student",
                            "teacher",
                            "parent",
                            "school",
                            "admin"
                        ]
                    },
                    "teacher": {
                        "properties": {
                            "city_id": {
                                "type": "integer"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "student": {
                        "properties": {
                            "grade_id": {
                                "type": "integer"
                            },
                            "city_id": {
                                "type": "integer"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "school": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "no": {
                                "type": "string",
                                "nullable": true
                            },
                            "city_id": {
                                "type": "integer"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "ValidationError": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "bearer",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}